move event code out of api index file
